+++ CHANGELOG.md
@@ -6,8 +6,7 @@
 - [js-web/vue-vite] improve structure for deployment of multiple apps
 - [chore] add fido2 spa sample in [js-node/expressjs], clean up ssr example from [js-node/wip/fido2]
 - [js-web/vue-vite] fix login & logout add loading spinner
-- In progress
-  - [js-web/vue-vite] cypress e2e testing (logout...)
+- [js-web/vue-vite] cypress e2e testing
 
 ### Version 0.6.5
 - [chore] update packages & cleanup & work on improving documentation